fix(economia): voce_extra_id deve essere BIGINT (id voci = timestamp ms)
integer out of range: gli id delle voci del computo sono int(now()*1000) ~1.8e12,
fuori dal range INTEGER. Migrazione 0014 ALTER a BIGINT. Questo bloccava
silenziosamente la creazione della voce extra preventivo dalle ore (errore
inghiottito dal try/except in _sync_voce_extra).

+++ backend/alembic/versions/0013_ore_extra_voce_computo.py
@@ -21,7 +21,7 @@ def _has_col(bind, table, col):
 def upgrade():
     bind = op.get_bind()
     if not _has_col(bind, "ore_extra", "voce_extra_id"):
-        op.add_column("ore_extra", sa.Column("voce_extra_id", sa.Integer(), nullable=True))
+        op.add_column("ore_extra", sa.Column("voce_extra_id", sa.BigInteger(), nullable=True))
 
 
 def downgrade():
